+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 See [Keep a Changelog](https://keepachangelog.com/) for details.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Upcoming] - Unreleased
+
+### Changed
+
+- Temperature now shows only one decimal place for better readability.
+
 ## [0.1.5] - 2025-08-27
 
 ### Added